requirements.txt
camelot-py[base]
PyPDF2<3.0